-Clientes_Dao = Se ha insertado función para actualizar el orden de los clientes y borrar el orden de los mismos. Con sus respectivas funciones en Clientes_Repository, Clientes_UseCase, Cliente_ViewModel.
-El campo delegación se ha cambiado de tipo Int a short, tanto en Clientes_Entity (implementado delegación en el cursor) como en Cliente.kt

-En OrdenarClientes_Activity, se ha implementado una función la cual recorre las posiciones de los clientes, para posteriormente, obtener el orden elegido por el usuario, mediante la función definida en Cliente_ViewModel.
Se ha agregado, dos botones, aceptar y borrar, con sus respectivos xml y su menú.

+++ app/src/main/java/com/advantys/brsmovilidaderp/Domain/Models/Cliente.kt
@@ -5,7 +5,7 @@ import java.time.format.DateTimeFormatter
 
 class Cliente (
     var numClientes: Int?=null,
-    var delegacion: Int?=null,
+    var delegacion: Short?=null,
     var nombre: String?=null,
     var razonSocial: String?=null,
     var direccion: String?=null,
